requirements.txt
Flask==2.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
peewee==3.14.10
PyMySQL==1.0.2
Werkzeug==2.1.0
flask-cors
cryptography
PyJWT